add new documentation and rearrange stories
