--- projects/accedia-angular-components/src/stories/BarChart.mdx
@@ -1,45 +1,0 @@
-import { Meta, Canvas, Controls } from "@storybook/blocks";
-import * as BarChartStories from "./BarChart.stories.ts";
-import GithubIcon from "./assets/github-mark.svg";
-
-<style>
-  {`
-    .github-icon {
-      margin-left: 0.25rem;
-    }
-  `}
-</style>
-
-# Bar chart
-
-The <strong>Bar Chart</strong> component enables the visualization of data using horizontal bars of dynamic lengths. 
-It offers customization options for component labels, animations, gaps between bars, and legend display.
-
-<Meta of={BarChartStories} />
-
-<Canvas of={BarChartStories.Medium} />
-
-<Controls />
-
-## Functionality
-
-<ul>
-  <li>Renders horizontal bars representing different components with their respective values.</li>
-  <li>Supports two types of component labels: `amount` for displaying absolute values and `percentage` for displaying values as percentages of the total.</li>
-  <li>Provides built-in animations for enhancing user experience when components are added or updated.</li>
-  <li>Offers customization options for adjusting the gap between bars, displaying component labels, and showing/hiding the legend.</li>
-</ul>
-
-## Usage
-
-Integrate the Bar Chart component into your Angular application by adding the `<acc-bar-chart>` element to your template. 
-Bind the `components` input property to an array of objects representing the components to be displayed in the bar chart. 
-Each object should contain a <strong>value</strong> property representing the component's value, 
-an optional <strong>color</strong> property specifying the bar color, and an optional <strong>description</strong> to be displayed in the legend.
-Leaving out the color property of a component will result in a default color being assigned to it. The 16 default colors are public and can be exported. 
-<br/><br/>You can customize the behaviour and appearance using input properties such as `componentLabelType` to specify the type of component labels to display, 
-`disableAnimations` to disable animations, and `gap` to set the spacing between bars.
-
-## Source code <img className="github-icon" src={GithubIcon} alt="github" width="20"/>
-
-[Bar chart component](https://github.com/histefanov/accedia-angular-components/tree/main/projects/accedia-angular-components/src/lib/components/bar-chart)
