--- projects/accedia-angular-components/src/stories/Button.example.stories.ts
@@ -1,51 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/angular';
-import Button from './button.component';
-
-// More on how to set up stories at: https://storybook.js.org/docs/angular/writing-stories/introduction
-const meta: Meta<Button> = {
-  title: 'Example/ExampleButton',
-  component: Button,
-  tags: ['autodocs'],
-  render: (args: Button) => ({
-    props: {
-      backgroundColor: null,
-      ...args,
-    },
-  }),
-  argTypes: {
-    backgroundColor: {
-      control: 'color',
-    },
-  },
-};
-
-export default meta;
-type Story = StoryObj<Button>;
-
-// More on writing stories with args: https://storybook.js.org/docs/angular/writing-stories/args
-export const PrimaryExampleButton: Story = {
-  args: {
-    primary: true,
-    label: 'Button',
-  },
-};
-
-export const SecondaryExampleButton: Story = {
-  args: {
-    label: 'Button',
-  },
-};
-
-export const Large: Story = {
-  args: {
-    size: 'large',
-    label: 'Button',
-  },
-};
-
-export const Small: Story = {
-  args: {
-    size: 'small',
-    label: 'Button',
-  },
-};
