--- projects/accedia-angular-components/src/stories/RadioGroup.mdx
@@ -1,11 +1,0 @@
-import { Meta, Canvas, Controls } from "@storybook/blocks";
-import * as RadioGroupStories from "./RadioGroup.stories.ts";
-import GithubIcon from "./assets/github-mark.svg";
-
-# Radio button
-
-<Meta of={RadioGroupStories} />
-
-<Canvas of={RadioGroupStories.Primary} />
-
-<Controls />
