--- projects/accedia-angular-components/src/stories/Tab.mdx
@@ -1,31 +1,0 @@
-import { Meta, Canvas, Controls } from "@storybook/blocks";
-import * as TabStories from "./Tab.stories.ts";
-import GithubIcon from "./assets/github-mark.svg";
-
-<style>
-  {`
-    .github-icon {
-      margin-left: 0.25rem;
-    }
-  `}
-</style>
-
-# Tab
-
-The <strong>Tab</strong> component is a fundamental building block of the tabbed navigation 
-system, designed to be used as a child component within the <strong>Tabs</strong> component. 
-It represents an individual tab with its label and associated content.
-
-For detailed information and usage examples, please refer to the <strong>Tabs</strong> component documentation. 
-It provides comprehensive information on implementing tab navigation and how to effectively utilize the 
-<strong>Tab</strong> component within the context of the entire tabbed navigation system.
-
-<Meta of={TabStories} />
-
-<Canvas of={TabStories.Primary} />
-
-<Controls />
-
-## Source code <img className="github-icon" src={GithubIcon} alt="github" width="20"/>
-
-[Tab component](https://github.com/histefanov/accedia-angular-components/tree/main/projects/accedia-angular-components/src/lib/components/tab)
