+++ projects/accedia-angular-components/src/stories/expandable-container.stories.ts
@@ -8,7 +8,7 @@ import { AdditionalInfoComponent } from '../lib/components/additional-info/addit
 
 // More on how to set up stories at: https://storybook.js.org/docs/angular/writing-stories/introduction
 const meta: Meta<ExpandableContainerComponent> = {
-    title: 'Example/ExpandableShowcaseComponent',
+    title: 'Components/ExpandableShowcaseComponent',
     component: ExpandableContainerComponent,
     tags: ['autodocs'],
     decorators: [
