+++ projects/accedia-angular-components/src/stories/paginator/Paginator.mdx
@@ -1,0 +1,41 @@
+import { Meta, Canvas, Controls } from "@storybook/blocks";
+import * as PaginatorStories from "./Paginator.stories.ts";
+import GithubIcon from "../assets/github-mark.svg";
+
+<style>
+  {`
+    .github-icon {
+      margin-left: 0.25rem;
+    }
+  `}
+</style>
+
+# Paginator
+
+The <strong>Paginator</strong> component provides pagination functionality for navigating through a collection of items. It enables users to control the number of items displayed per page and navigate through pages efficiently.
+
+<Meta of={PaginatorStories} />
+
+<Canvas of={PaginatorStories.Primary} />
+
+<Controls />
+
+## Functionality
+
+<ul>
+  <li>Displays pagination controls such as previous, next, first, and last buttons.</li>
+  <li>Allows customization of the number of visible page buttons.</li>
+  <li>Supports customization of page size and provides options for selecting different page sizes.</li>
+  <li>Emits events when the page changes, providing the updated page index, page size, and total number of items.</li>
+</ul>
+
+## Usage
+
+The Paginator component is integrated into your Angular application by adding the `<acc-paginator>` element to your template.
+Configure its inputs to customize its behavior, such as setting the total number of items using `itemsLength`, specifying the current page index
+with `pageIndex`, defining the number of visible page buttons with `pageButtonsCount`, etc. (please refer to the component's <strong>Controls</strong> for further information).<br/>
+Implement the `(pageChange)` event handler to capture page change events and update the data accordingly.
+
+## Source code <img className="github-icon" src={GithubIcon} alt="github" width="20"/>
+
+[Paginator component](https://github.com/histefanov/accedia-angular-components/tree/main/projects/accedia-angular-components/src/lib/components/paginator)
