--- projects/accedia-angular-components/src/stories/progressBar.mdx
@@ -1,38 +1,0 @@
-{/* ProgressBarDocs.mdx */}
-
-import { Meta, Primary, Controls, Story, Canvas } from '@storybook/blocks';
-import * as ProgressBar from './progress-bar.stories';
-
-<Meta of={ProgressBar} />
-
-# ProgressBar
-
-A simple progress bar that has a progress animation that can be tweaked depending on the use case.
-
-<Primary />
-
-<Controls />
-
-{/* ## Stories
-
-### Primary */}
-
-{/* <Story of={ProgressBar.Primary} /> */}
-
-{/* <Canvas
-  additionalActions={[
-    {
-      title: 'Open in GitHub',
-      onClick: () => {
-        window.open(
-          'https://github.com/storybookjs/storybook/blob/next/code/ui/blocks/src/examples/Button.stories.tsx',
-          '_blank'
-        );
-      },
-    },
-  ]}
-  of={ProgressBar.Primary}
-/> */}
-
-
-{/* ... */}
